feat: stability, fast restart, html rendering
* disable yt-dlp, saving video url to a file for manual processing later
* render html for audio and video
* skip download if file exists and sizes match
* if `ignore_me` file is found inside post folder. skip processing
* pagination offset is stored in `last.offset` file for fast resume
* better page walk, retries and tracebacks

+++ boosty_downloader/src/infrastructure/file_downloader.py
@@ -133,12 +133,17 @@ async def download_file(
             if ext is not None:
                 file_path = file_path.with_suffix(ext)
 
-        total_downloaded = 0
 
+        if file_path.exists() and file_path.stat().st_size == response.content_length:
+            #print(f'File exists and size is right, skip download: {file_path}')
+            return file_path
+
+        total_downloaded = 0
         async with aiofiles.open(file_path, mode='wb') as file:
             total_size = response.content_length
 
             try:
+                # sometimes fails with ContentLengthError: 400 Not enough data to satisfy content length header
                 async for chunk in response.content.iter_chunked(
                     dl_config.chunk_size_bytes
                 ):
